+++ src/collective/cover/tests/test_richtext_tile.py
@@ -17,7 +17,7 @@ class RichTextTileTestCase(unittest.TestCase):
         self.portal = self.layer['portal']
         self.request = self.layer['request']
         self.tile = self.portal.restrictedTraverse(
-            '@@%s/%s' % ('collective.cover.richtext', 'test-richtext-tile'))
+            '@@{0}/{1}'.format('collective.cover.richtext', 'test-richtext-tile'))
 
     def test_interface(self):
         self.assertTrue(IPersistentCoverTile.implementedBy(RichTextTile))